feat(UseCase): finish useCase implementation && feat(JWT/GoogleAuth): add auth with jwt and start the google auth implementation

+++ src/main/java/com/java/pink_closet/dto/orderItem/response/OrderItemDetailedResponse.java
@@ -2,12 +2,14 @@ package com.java.pink_closet.dto.orderItem.response;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class OrderItemDetailedResponse {
     private Long id;
     private Long orderId;
     private Long productId;
     private String productName;
-    private Double productPrice;
+    private BigDecimal productPrice;
     private Integer quantity;
 }
